+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Campaign Changelog
 
+## 1.0.1 - 2018-05-03
+### Fixed
+- Fixed pending contacts source column
+
 ## 1.0.0 - 2018-05-02
 - Stable release